// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type CoursesDocumentDataSlicesSlice = never;

/**
 * Content for Courses documents
 */
interface CoursesDocumentData {
  /**
   * Course Banner field in *Courses*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: courses.banner
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  banner: prismic.ImageField<never>;

  /**
   * Nome do Curso field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Nome do Curso
   * - **API ID Path**: courses.curso
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  curso: prismic.KeyTextField;

  /**
   * Descrição field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Descrição do Curso
   * - **API ID Path**: courses.descricao
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  descricao: prismic.KeyTextField;

  /**
   * Descrição Extensa field in *Courses*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Descrição detalhada
   * - **API ID Path**: courses.descricao_extensa
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  descricao_extensa: prismic.RichTextField;

  /**
   * Mercado de Trabalho field in *Courses*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Mercado de Trabalho
   * - **API ID Path**: courses.mercado_de_trabalho
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  mercado_de_trabalho: prismic.RichTextField;

  /**
   * Infraestrutura field in *Courses*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Infraestrutura
   * - **API ID Path**: courses.infraestrutura
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  infraestrutura: prismic.RichTextField;

  /**
   * Disiplinas field in *Courses*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: Disciplinas
   * - **API ID Path**: courses.disiplinas
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#rich-text-title
   */
  disiplinas: prismic.RichTextField;

  /**
   * Semestres field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Semestres
   * - **API ID Path**: courses.semestres
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  semestres: prismic.KeyTextField;

  /**
   * Modalidade field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Modalidade
   * - **API ID Path**: courses.modalidade
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  modalidade: prismic.KeyTextField;

  /**
   * Slice Zone field in *Courses*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: courses.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  slices: prismic.SliceZone<CoursesDocumentDataSlicesSlice> /**
   * Meta Title field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: courses.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Courses*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: courses.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Courses*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: courses.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  meta_image: prismic.ImageField<never>;
}